move SpanFilterBuilder and collaborators to otel package
