+++ splunk-otel-android/src/main/java/com/splunk/rum/StandardAttributes.java
@@ -18,6 +18,7 @@ package com.splunk.rum;
 
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
+import io.opentelemetry.rum.internal.SpanFilterBuilder;
 import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
 
 /**
