+++ splunk-otel-android/src/main/java/io/opentelemetry/rum/internal/ModifiedSpanData.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.splunk.rum;
+package io.opentelemetry.rum.internal;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.sdk.trace.data.DelegatingSpanData;
